requirements.txt
absl-py==1.2.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
cachetools==5.2.0
charset-normalizer==2.1.0
coloredlogs==15.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
executing==0.9.0
flatbuffers==2.0.7
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
humanfriendly==10.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mpmath==1.2.1
mysql-connector-python==8.0.29
numpy==1.23.1
oauthlib==3.2.0
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
packaging==21.3
pafy==0.5.5
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.3.0
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.0
wrapt==1.14.1
youtube-dl==2020.12.2
